Code refactored

+++ src/main/java/com/ltap/usermanagement/repository/UserRepo.java
@@ -1,10 +1,8 @@
 package com.ltap.usermanagement.repository;
 
 import com.ltap.usermanagement.entities.UserInfo;
-
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
-
 @Repository
-public interface UserRepo extends CrudRepository<UserInfo, Integer> {}
+public interface UserRepo extends CrudRepository<UserInfo, Long> {}
